move docs repo to separate file, refactor tests
